backend/requirements.txt
Django==3.2.15
django-cors-headers==3.6.0
djangorestframework==3.13.1

